model/requirements.txt
apache-flink-statefun==2.2.0
flask==1.1.1
gunicorn==20.0.4
